+++ sea-orm-codegen/src/entity/column.rs
@@ -1,8 +1,8 @@
-use crate::util::escape_rust_keyword;
+use crate::{util::escape_rust_keyword, DateTimeCrate};
 use heck::{CamelCase, SnakeCase};
 use proc_macro2::{Ident, TokenStream};
 use quote::{format_ident, quote};
-use sea_query::{ColumnDef, ColumnSpec, ColumnType};
+use sea_query::{BlobSize, ColumnDef, ColumnSpec, ColumnType};
 use std::fmt::Write as FmtWrite;
 
 #[derive(Clone, Debug)]
@@ -27,7 +27,7 @@ impl Column {
         self.name.to_snake_case() == self.name
     }
 
-    pub fn get_rs_type(&self) -> TokenStream {
+    pub fn get_rs_type(&self, date_time_crate: &DateTimeCrate) -> TokenStream {
         #[allow(unreachable_patterns)]
         let ident: TokenStream = match &self.col_type {
             ColumnType::Char(_)
@@ -45,11 +45,27 @@ impl Column {
             ColumnType::Float(_) => "f32".to_owned(),
             ColumnType::Double(_) => "f64".to_owned(),
             ColumnType::Json | ColumnType::JsonBinary => "Json".to_owned(),
-            ColumnType::Date => "Date".to_owned(),
-            ColumnType::Time(_) => "Time".to_owned(),
-            ColumnType::DateTime(_) => "DateTime".to_owned(),
-            ColumnType::Timestamp(_) => "DateTimeUtc".to_owned(),
-            ColumnType::TimestampWithTimeZone(_) => "DateTimeWithTimeZone".to_owned(),
+            ColumnType::Date => match date_time_crate {
+                DateTimeCrate::Chrono => "Date".to_owned(),
+                DateTimeCrate::Time => "TimeDate".to_owned(),
+            },
+            ColumnType::Time(_) => match date_time_crate {
+                DateTimeCrate::Chrono => "Time".to_owned(),
+                DateTimeCrate::Time => "TimeTime".to_owned(),
+            },
+            ColumnType::DateTime(_) => match date_time_crate {
+                DateTimeCrate::Chrono => "DateTime".to_owned(),
+                DateTimeCrate::Time => "TimeDateTime".to_owned(),
+            },
+            ColumnType::Timestamp(_) => match date_time_crate {
+                DateTimeCrate::Chrono => "DateTimeUtc".to_owned(),
+                // ColumnType::Timpestamp(_) => time::PrimitiveDateTime: https://docs.rs/sqlx/0.3.5/sqlx/postgres/types/index.html#time
+                DateTimeCrate::Time => "TimeDateTime".to_owned(),
+            },
+            ColumnType::TimestampWithTimeZone(_) => match date_time_crate {
+                DateTimeCrate::Chrono => "DateTimeWithTimeZone".to_owned(),
+                DateTimeCrate::Time => "TimeDateTimeWithTimeZone".to_owned(),
+            },
             ColumnType::Decimal(_) | ColumnType::Money(_) => "Decimal".to_owned(),
             ColumnType::Uuid => "Uuid".to_owned(),
             ColumnType::Binary(_) => "Vec<u8>".to_owned(),
@@ -112,7 +128,10 @@ impl Column {
             }
             ColumnType::Time(_) => quote! { ColumnType::Time.def() },
             ColumnType::Date => quote! { ColumnType::Date.def() },
-            ColumnType::Binary(_) => quote! { ColumnType::Binary.def() },
+            ColumnType::Binary(BlobSize::Blob(_)) => quote! { ColumnType::Binary.def() },
+            ColumnType::Binary(BlobSize::Tiny) => quote! { ColumnType::TinyBinary.def() },
+            ColumnType::Binary(BlobSize::Medium) => quote! { ColumnType::MediumBinary.def() },
+            ColumnType::Binary(BlobSize::Long) => quote! { ColumnType::LongBinary.def() },
             ColumnType::Boolean => quote! { ColumnType::Boolean.def() },
             ColumnType::Money(s) => match s {
                 Some((s1, s2)) => quote! { ColumnType::Money(Some((#s1, #s2))).def() },
@@ -147,7 +166,10 @@ impl Column {
 
     pub fn get_info(&self) -> String {
         let mut info = String::new();
-        let type_info = self.get_rs_type().to_string().replace(' ', "");
+        let type_info = self
+            .get_rs_type(&DateTimeCrate::Chrono)
+            .to_string()
+            .replace(' ', "");
         let col_info = self.col_info();
         write!(
             &mut info,
@@ -210,10 +232,10 @@ impl From<&ColumnDef> for Column {
 
 #[cfg(test)]
 mod tests {
-    use crate::Column;
+    use crate::{Column, DateTimeCrate};
     use proc_macro2::TokenStream;
     use quote::quote;
-    use sea_query::{Alias, ColumnDef, ColumnType, SeaRc};
+    use sea_query::{Alias, BlobSize, ColumnDef, ColumnType, SeaRc};
 
     fn setup() -> Vec<Column> {
         macro_rules! make_col {
@@ -243,7 +265,7 @@ mod tests {
             make_col!("CakeFillingId", ColumnType::BigUnsigned(Some(12))),
             make_col!("cake-filling-id", ColumnType::Float(None)),
             make_col!("CAKE_FILLING_ID", ColumnType::Double(None)),
-            make_col!("CAKE-FILLING-ID", ColumnType::Binary(None)),
+            make_col!("CAKE-FILLING-ID", ColumnType::Binary(BlobSize::Blob(None))),
             make_col!("CAKE", ColumnType::Boolean),
             make_col!("date", ColumnType::Date),
             make_col!("time", ColumnType::Time(None)),
@@ -312,8 +334,9 @@ mod tests {
     }
 
     #[test]
-    fn test_get_rs_type() {
+    fn test_get_rs_type_with_chrono() {
         let columns = setup();
+        let chrono_crate = DateTimeCrate::Chrono;
         let rs_types = vec![
             "String",
             "String",
@@ -339,11 +362,56 @@ mod tests {
             let rs_type: TokenStream = rs_type.parse().unwrap();
 
             col.not_null = true;
-            assert_eq!(col.get_rs_type().to_string(), quote!(#rs_type).to_string());
+            assert_eq!(
+                col.get_rs_type(&chrono_crate).to_string(),
+                quote!(#rs_type).to_string()
+            );
 
             col.not_null = false;
             assert_eq!(
-                col.get_rs_type().to_string(),
+                col.get_rs_type(&chrono_crate).to_string(),
+                quote!(Option<#rs_type>).to_string()
+            );
+        }
+    }
+
+    #[test]
+    fn test_get_rs_type_with_time() {
+        let columns = setup();
+        let time_crate = DateTimeCrate::Time;
+        let rs_types = vec![
+            "String",
+            "String",
+            "i8",
+            "u8",
+            "i16",
+            "u16",
+            "i32",
+            "u32",
+            "i64",
+            "u64",
+            "f32",
+            "f64",
+            "Vec<u8>",
+            "bool",
+            "TimeDate",
+            "TimeTime",
+            "TimeDateTime",
+            "TimeDateTime",
+            "TimeDateTimeWithTimeZone",
+        ];
+        for (mut col, rs_type) in columns.into_iter().zip(rs_types) {
+            let rs_type: TokenStream = rs_type.parse().unwrap();
+
+            col.not_null = true;
+            assert_eq!(
+                col.get_rs_type(&time_crate).to_string(),
+                quote!(#rs_type).to_string()
+            );
+
+            col.not_null = false;
+            assert_eq!(
+                col.get_rs_type(&time_crate).to_string(),
                 quote!(Option<#rs_type>).to_string()
             );
         }