+++ sea-orm-codegen/src/entity/writer.rs
@@ -29,6 +29,20 @@ pub enum WithSerde {
     Both,
 }
 
+#[derive(Debug)]
+pub enum DateTimeCrate {
+    Chrono,
+    Time,
+}
+
+#[derive(Debug)]
+pub struct EntityWriterContext {
+    pub(crate) expanded_format: bool,
+    pub(crate) with_serde: WithSerde,
+    pub(crate) date_time_crate: DateTimeCrate,
+    pub(crate) schema_name: Option<String>,
+}
+
 impl WithSerde {
     pub fn extra_derive(&self) -> TokenStream {
         let mut extra_derive = match self {
@@ -79,29 +93,35 @@ impl FromStr for WithSerde {
     }
 }
 
-impl EntityWriter {
-    pub fn generate(
-        self,
+impl EntityWriterContext {
+    pub fn new(
         expanded_format: bool,
         with_serde: WithSerde,
+        date_time_crate: DateTimeCrate,
         schema_name: Option<String>,
-    ) -> WriterOutput {
+    ) -> Self {
+        Self {
+            expanded_format,
+            with_serde,
+            date_time_crate,
+            schema_name,
+        }
+    }
+}
+
+impl EntityWriter {
+    pub fn generate(self, context: &EntityWriterContext) -> WriterOutput {
         let mut files = Vec::new();
-        files.extend(self.write_entities(expanded_format, &with_serde, schema_name));
+        files.extend(self.write_entities(context));
         files.push(self.write_mod());
         files.push(self.write_prelude());
         if !self.enums.is_empty() {
-            files.push(self.write_sea_orm_active_enums(&with_serde));
+            files.push(self.write_sea_orm_active_enums(&context.with_serde));
         }
         WriterOutput { files }
     }
 
-    pub fn write_entities(
-        &self,
-        expanded_format: bool,
-        with_serde: &WithSerde,
-        schema_name: Option<String>,
-    ) -> Vec<OutputFile> {
+    pub fn write_entities(&self, context: &EntityWriterContext) -> Vec<OutputFile> {
         self.entities
             .iter()
             .map(|entity| {
@@ -119,10 +139,20 @@ impl EntityWriter {
 
                 let mut lines = Vec::new();
                 Self::write_doc_comment(&mut lines);
-                let code_blocks = if expanded_format {
-                    Self::gen_expanded_code_blocks(entity, with_serde, &schema_name)
+                let code_blocks = if context.expanded_format {
+                    Self::gen_expanded_code_blocks(
+                        entity,
+                        &context.with_serde,
+                        &context.date_time_crate,
+                        &context.schema_name,
+                    )
                 } else {
-                    Self::gen_compact_code_blocks(entity, with_serde, &schema_name)
+                    Self::gen_compact_code_blocks(
+                        entity,
+                        &context.with_serde,
+                        &context.date_time_crate,
+                        &context.schema_name,
+                    )
                 };
                 Self::write(&mut lines, code_blocks);
                 OutputFile {
@@ -209,6 +239,7 @@ impl EntityWriter {
     pub fn gen_expanded_code_blocks(
         entity: &Entity,
         with_serde: &WithSerde,
+        date_time_crate: &DateTimeCrate,
         schema_name: &Option<String>,
     ) -> Vec<TokenStream> {
         let mut imports = Self::gen_import(with_serde);
@@ -217,10 +248,10 @@ impl EntityWriter {
             imports,
             Self::gen_entity_struct(),
             Self::gen_impl_entity_name(entity, schema_name),
-            Self::gen_model_struct(entity, with_serde),
+            Self::gen_model_struct(entity, with_serde, date_time_crate),
             Self::gen_column_enum(entity),
             Self::gen_primary_key_enum(entity),
-            Self::gen_impl_primary_key(entity),
+            Self::gen_impl_primary_key(entity, date_time_crate),
             Self::gen_relation_enum(entity),
             Self::gen_impl_column_trait(entity),
             Self::gen_impl_relation_trait(entity),
@@ -234,13 +265,14 @@ impl EntityWriter {
     pub fn gen_compact_code_blocks(
         entity: &Entity,
         with_serde: &WithSerde,
+        date_time_crate: &DateTimeCrate,
         schema_name: &Option<String>,
     ) -> Vec<TokenStream> {
         let mut imports = Self::gen_import(with_serde);
         imports.extend(Self::gen_import_active_enum(entity));
         let mut code_blocks = vec![
             imports,
-            Self::gen_compact_model_struct(entity, with_serde, schema_name),
+            Self::gen_compact_model_struct(entity, with_serde, date_time_crate, schema_name),
         ];
         let relation_defs = if entity.get_relation_enum_name().is_empty() {
             vec![
@@ -332,9 +364,13 @@ impl EntityWriter {
             })
     }
 
-    pub fn gen_model_struct(entity: &Entity, with_serde: &WithSerde) -> TokenStream {
+    pub fn gen_model_struct(
+        entity: &Entity,
+        with_serde: &WithSerde,
+        date_time_crate: &DateTimeCrate,
+    ) -> TokenStream {
         let column_names_snake_case = entity.get_column_names_snake_case();
-        let column_rs_types = entity.get_column_rs_types();
+        let column_rs_types = entity.get_column_rs_types(date_time_crate);
 
         let extra_derive = with_serde.extra_derive();
 
@@ -377,9 +413,9 @@ impl EntityWriter {
         }
     }
 
-    pub fn gen_impl_primary_key(entity: &Entity) -> TokenStream {
+    pub fn gen_impl_primary_key(entity: &Entity, date_time_crate: &DateTimeCrate) -> TokenStream {
         let primary_key_auto_increment = entity.get_primary_key_auto_increment();
-        let value_type = entity.get_primary_key_rs_type();
+        let value_type = entity.get_primary_key_rs_type(date_time_crate);
         quote! {
             impl PrimaryKeyTrait for PrimaryKey {
                 type ValueType = #value_type;
@@ -515,11 +551,12 @@ impl EntityWriter {
     pub fn gen_compact_model_struct(
         entity: &Entity,
         with_serde: &WithSerde,
+        date_time_crate: &DateTimeCrate,
         schema_name: &Option<String>,
     ) -> TokenStream {
         let table_name = entity.table_name.as_str();
         let column_names_snake_case = entity.get_column_names_snake_case();
-        let column_rs_types = entity.get_column_rs_types();
+        let column_rs_types = entity.get_column_rs_types(date_time_crate);
         let primary_keys: Vec<String> = entity
             .primary_keys
             .iter()
@@ -619,8 +656,8 @@ impl EntityWriter {
 #[cfg(test)]
 mod tests {
     use crate::{
-        Column, ConjunctRelation, Entity, EntityWriter, PrimaryKey, Relation, RelationType,
-        WithSerde,
+        Column, ConjunctRelation, DateTimeCrate, Entity, EntityWriter, PrimaryKey, Relation,
+        RelationType, WithSerde,
     };
     use pretty_assertions::assert_eq;
     use proc_macro2::TokenStream;
@@ -1046,6 +1083,7 @@ mod tests {
                 EntityWriter::gen_expanded_code_blocks(
                     entity,
                     &crate::WithSerde::None,
+                    &crate::DateTimeCrate::Chrono,
                     &Some("public".to_owned())
                 )
                 .into_iter()
@@ -1061,6 +1099,7 @@ mod tests {
                 EntityWriter::gen_expanded_code_blocks(
                     entity,
                     &crate::WithSerde::None,
+                    &crate::DateTimeCrate::Chrono,
                     &Some("schema_name".to_owned())
                 )
                 .into_iter()
@@ -1115,6 +1154,7 @@ mod tests {
                 EntityWriter::gen_compact_code_blocks(
                     entity,
                     &crate::WithSerde::None,
+                    &crate::DateTimeCrate::Chrono,
                     &Some("public".to_owned())
                 )
                 .into_iter()
@@ -1130,6 +1170,7 @@ mod tests {
                 EntityWriter::gen_compact_code_blocks(
                     entity,
                     &crate::WithSerde::None,
+                    &crate::DateTimeCrate::Chrono,
                     &Some("schema_name".to_owned())
                 )
                 .into_iter()
@@ -1234,7 +1275,7 @@ mod tests {
     fn assert_serde_variant_results(
         cake_entity: &Entity,
         entity_serde_variant: &(String, WithSerde, Option<String>),
-        generator: Box<dyn Fn(&Entity, &WithSerde, &Option<String>) -> Vec<TokenStream>>,
+        generator: Box<dyn Fn(&Entity, &WithSerde, &DateTimeCrate, &Option<String>) -> Vec<TokenStream>>,
     ) -> io::Result<()> {
         let mut reader = BufReader::new(entity_serde_variant.0.as_bytes());
         let mut lines: Vec<String> = Vec::new();
@@ -1251,6 +1292,7 @@ mod tests {
         let generated = generator(
             cake_entity,
             &entity_serde_variant.1,
+            &DateTimeCrate::Chrono
             &entity_serde_variant.2,
         )
         .into_iter()