+++ src/database/stream/metric.rs
@@ -1,8 +1,8 @@
-use std::{time::Duration, pin::Pin, task::Poll};
+use std::{pin::Pin, task::Poll, time::Duration};
 
 use futures::Stream;
 
-use crate::{QueryResult, DbErr, Statement};
+use crate::{DbErr, QueryResult, Statement};
 
 pub(crate) struct MetricStream<'a> {
     metric_callback: &'a Option<crate::metric::Callback>,
@@ -12,7 +12,13 @@ pub(crate) struct MetricStream<'a> {
 }
 
 impl<'a> MetricStream<'a> {
-    pub(crate) fn new<S>(metric_callback: &'a Option<crate::metric::Callback>, stmt: &'a Statement, elapsed: Option<Duration>, stream: S) -> Self
+    #[allow(dead_code)]
+    pub(crate) fn new<S>(
+        metric_callback: &'a Option<crate::metric::Callback>,
+        stmt: &'a Statement,
+        elapsed: Option<Duration>,
+        stream: S,
+    ) -> Self
     where
         S: Stream<Item = Result<QueryResult, DbErr>> + 'a + Send,
     {
@@ -33,7 +39,10 @@ impl<'a> Stream for MetricStream<'a> {
         cx: &mut std::task::Context<'_>,
     ) -> Poll<Option<Self::Item>> {
         let this = self.get_mut();
-        let _start = this.metric_callback.is_some().then(std::time::SystemTime::now);
+        let _start = this
+            .metric_callback
+            .is_some()
+            .then(std::time::SystemTime::now);
         let res = Pin::new(&mut this.stream).poll_next(cx);
         if let (Some(_start), Some(elapsed)) = (_start, &mut this.elapsed) {
             *elapsed += _start.elapsed().unwrap_or_default();
@@ -46,7 +55,7 @@ impl<'a> Drop for MetricStream<'a> {
     fn drop(&mut self) {
         if let (Some(callback), Some(elapsed)) = (self.metric_callback.as_deref(), self.elapsed) {
             let info = crate::metric::Info {
-                elapsed: elapsed,
+                elapsed,
                 statement: self.stmt,
                 failed: false,
             };