+++ src/database/stream/query.rs
@@ -1,6 +1,6 @@
-#![allow(missing_docs)]
+#![allow(missing_docs, unreachable_code, unused_variables)]
 
-use std::{pin::Pin, task::Poll, time::SystemTime};
+use std::{pin::Pin, task::Poll};
 
 #[cfg(feature = "mock")]
 use std::sync::Arc;
@@ -126,40 +126,43 @@ impl QueryStream {
                 #[cfg(feature = "sqlx-mysql")]
                 InnerConnection::MySql(c) => {
                     let query = crate::driver::sqlx_mysql::sqlx_query(stmt);
-                    let _start = _metric_callback.is_some().then(SystemTime::now);
-                    let stream = c.fetch(query)
-                                .map_ok(Into::into)
-                                .map_err(crate::sqlx_error_to_query_err);
+                    let _start = _metric_callback.is_some().then(std::time::SystemTime::now);
+                    let stream = c
+                        .fetch(query)
+                        .map_ok(Into::into)
+                        .map_err(crate::sqlx_error_to_query_err);
                     let elapsed = _start.map(|s| s.elapsed().unwrap_or_default());
                     MetricStream::new(_metric_callback, stmt, elapsed, stream)
                 }
                 #[cfg(feature = "sqlx-postgres")]
                 InnerConnection::Postgres(c) => {
                     let query = crate::driver::sqlx_postgres::sqlx_query(stmt);
-                    let _start = _metric_callback.is_some().then(SystemTime::now);
-                    let stream = c.fetch(query)
-                                .map_ok(Into::into)
-                                .map_err(crate::sqlx_error_to_query_err);
+                    let _start = _metric_callback.is_some().then(std::time::SystemTime::now);
+                    let stream = c
+                        .fetch(query)
+                        .map_ok(Into::into)
+                        .map_err(crate::sqlx_error_to_query_err);
                     let elapsed = _start.map(|s| s.elapsed().unwrap_or_default());
                     MetricStream::new(_metric_callback, stmt, elapsed, stream)
                 }
                 #[cfg(feature = "sqlx-sqlite")]
                 InnerConnection::Sqlite(c) => {
                     let query = crate::driver::sqlx_sqlite::sqlx_query(stmt);
-                    let _start = _metric_callback.is_some().then(SystemTime::now);
-                    let stream = c.fetch(query)
-                                .map_ok(Into::into)
-                                .map_err(crate::sqlx_error_to_query_err);
+                    let _start = _metric_callback.is_some().then(std::time::SystemTime::now);
+                    let stream = c
+                        .fetch(query)
+                        .map_ok(Into::into)
+                        .map_err(crate::sqlx_error_to_query_err);
                     let elapsed = _start.map(|s| s.elapsed().unwrap_or_default());
                     MetricStream::new(_metric_callback, stmt, elapsed, stream)
                 }
                 #[cfg(feature = "mock")]
                 InnerConnection::Mock(c) => {
-                    let _start = _metric_callback.is_some().then(SystemTime::now);
+                    let _start = _metric_callback.is_some().then(std::time::SystemTime::now);
                     let stream = c.fetch(stmt);
                     let elapsed = _start.map(|s| s.elapsed().unwrap_or_default());
                     MetricStream::new(_metric_callback, stmt, elapsed, stream)
-                },
+                }
                 #[allow(unreachable_patterns)]
                 _ => unreachable!(),
             },