+++ src/lib.rs
@@ -33,6 +33,7 @@
 //! Join our Discord server to chat with others in the SeaQL community!
 //!
 //! + [Getting Started](https://www.sea-ql.org/SeaORM/docs/index)
+//! + [Step-by-step Tutorials](https://www.sea-ql.org/sea-orm-tutorial/)
 //! + [Usage Example](https://github.com/SeaQL/sea-orm/tree/master/examples/basic)
 //!
 //! Integration examples
@@ -329,8 +330,9 @@ pub use schema::*;
 #[cfg(feature = "macros")]
 pub use sea_orm_macros::{
     DeriveActiveEnum, DeriveActiveModel, DeriveActiveModelBehavior, DeriveColumn,
-    DeriveCustomColumn, DeriveEntity, DeriveEntityModel, DeriveIntoActiveModel, DeriveModel,
-    DerivePrimaryKey, DeriveRelation, FromQueryResult,
+    DeriveCustomColumn, DeriveEntity, DeriveEntityModel, DeriveIntoActiveModel,
+    DeriveMigrationName, DeriveModel, DerivePrimaryKey, DeriveRelation, FromJsonQueryResult,
+    FromQueryResult,
 };
 
 pub use sea_query;