+++ tests/timestamp_tests.rs
@@ -2,7 +2,6 @@ pub mod common;
 pub use common::{features::*, setup::*, TestContext};
 use pretty_assertions::assert_eq;
 use sea_orm::{entity::prelude::*, DatabaseConnection, IntoActiveModel};
-use serde_json::json;
 
 #[sea_orm_macros::test]
 #[cfg(any(
@@ -39,17 +38,17 @@ pub async fn create_applog(db: &DatabaseConnection) -> Result<(), DbErr> {
     #[cfg(feature = "sqlx-sqlite")]
     assert_eq!(
         Applog::find().into_json().one(db).await?,
-        Some(json!({
+        Some(serde_json::json!({
             "id": 1,
             "action": "Testing",
             "json": r#""HI""#,
-            "created_at": "2021-09-17 09:50:20",
+            "created_at": "2021-09-17T17:50:20+08:00",
         }))
     );
     #[cfg(feature = "sqlx-mysql")]
     assert_eq!(
         Applog::find().into_json().one(db).await?,
-        Some(json!({
+        Some(serde_json::json!({
             "id": 1,
             "action": "Testing",
             "json": "HI",
@@ -59,7 +58,7 @@ pub async fn create_applog(db: &DatabaseConnection) -> Result<(), DbErr> {
     #[cfg(feature = "sqlx-postgres")]
     assert_eq!(
         Applog::find().into_json().one(db).await?,
-        Some(json!({
+        Some(serde_json::json!({
             "id": 1,
             "action": "Testing",
             "json": "HI",
@@ -88,17 +87,17 @@ pub async fn create_satellites_log(db: &DatabaseConnection) -> Result<(), DbErr>
     #[cfg(feature = "sqlx-sqlite")]
     assert_eq!(
         Satellite::find().into_json().one(db).await?,
-        Some(json!({
+        Some(serde_json::json!({
             "id": 1,
             "satellite_name": "Sea-00001-2022",
-            "launch_date": "2022-01-07 12:11:23",
-            "deployment_date": "2022-01-07 12:11:23",
+            "launch_date": "2022-01-07T12:11:23+00:00",
+            "deployment_date": "2022-01-07T12:11:23Z".parse::<DateTimeLocal>().unwrap(),
         }))
     );
     #[cfg(feature = "sqlx-mysql")]
     assert_eq!(
         Satellite::find().into_json().one(db).await?,
-        Some(json!({
+        Some(serde_json::json!({
             "id": 1,
             "satellite_name": "Sea-00001-2022",
             "launch_date": "2022-01-07T12:11:23Z",
@@ -108,7 +107,7 @@ pub async fn create_satellites_log(db: &DatabaseConnection) -> Result<(), DbErr>
     #[cfg(feature = "sqlx-postgres")]
     assert_eq!(
         Satellite::find().into_json().one(db).await?,
-        Some(json!({
+        Some(serde_json::json!({
             "id": 1,
             "satellite_name": "Sea-00001-2022",
             "launch_date": "2022-01-07T12:11:23+00:00",